flink/q2-job/requirements.txt
apache-flink==2.0.0
numpy
pillow

